Conflicts:
	railo-java/railo-core/src/railo/runtime/Info.ini
	railo-java/railo-core/src/resource/context/railo-context.ra

+++ railo-java/railo-core/src/railo/runtime/tag/StoredProc.java
@@ -609,8 +609,8 @@ public class StoredProc extends BodyTagTryCatchFinallySupport {
 	/**
 	 * @param timeout the timeout to set
 	 */
-	public void setTimeout(int timeout) {
-		this.timeout = timeout;
+	public void setTimeout(double timeout) {
+		this.timeout = (int) timeout;
 	}
